Add files via upload
House Price Prediction using Linear Regression
The housing market is influenced by various factors, including the size of the house, the number of bedrooms and bathrooms, location, and more. This project focuses on building a simple yet effective linear regression model that uses square footage and the count of bedrooms and bathrooms as predictive features for estimating house prices.